Reduce java utils and date usages
